move Random from System to Util
