+++ config/usa/arm9/delinks.txt
@@ -74,7 +74,7 @@ src/Memory/SafeAllocator.cpp:
     complete
     .text start:0x0203246c end:0x020328bc
 
-src/System/Random.c:
+src/Util/Random.cpp:
 	complete
 	.text start:0x020741d0 end:0x020744a8
